osdep_get_cwd: linux: recurse parents until one yields
This helps when a command is running as a different user, it is common
to be able to see its parents, but not its cwd.  It may make sense, in
this situation, to present the parent's path.  e.g. if a user runs a set
of commands like:

$ cd /foo/bar
$ sudo hang-for-a-bit

.. it could be expected that get_cwd returns "/foo/bar", not NULL, as
previously.  sudo's cwd is not available, as the process itself is
running as root, but its parent process is, and its parent (the shell)'s
cwd is available to us.